[chore] [exporterhelper] Remove redundant persistent_storage files
After the persistentStorage is consolidated persistentQueue, there is no reason to separate code in separate files
